refactor: auth 컴포넌트를 service에서 support 패키지로 이동
- 핵심 도메인 로직이 아니면서 인증 흐름을 보조하는 역할이므로 support 패키지로 이동한다.
