+++ backend/src/main/java/com/yagubogu/auth/gateway/AuthGateway.java
@@ -1,4 +1,4 @@
-package com.yagubogu.auth.client;
+package com.yagubogu.auth.gateway;
 
 import com.yagubogu.auth.dto.LoginRequest;
 import com.yagubogu.auth.dto.AuthResponse;
