+++ backend/src/main/java/com/yagubogu/auth/gateway/LocalAuthGateway.java
@@ -1,4 +1,4 @@
-package com.yagubogu.auth.client;
+package com.yagubogu.auth.gateway;
 
 import com.yagubogu.auth.config.GoogleAuthProperties;
 import com.yagubogu.auth.dto.AuthResponse;
