+++ backend/src/main/java/com/yagubogu/auth/support/GoogleAuthValidator.java
@@ -1,4 +1,4 @@
-package com.yagubogu.auth.service;
+package com.yagubogu.auth.support;
 
 import com.yagubogu.auth.dto.GoogleAuthResponse;
 import com.yagubogu.auth.exception.InvalidTokenException;
