+++ backend/src/main/java/com/yagubogu/auth/support/JwtProvider.java
@@ -1,4 +1,4 @@
-package com.yagubogu.auth.service;
+package com.yagubogu.auth.support;
 
 
 import com.auth0.jwt.JWT;
