+++ backend/src/test/java/com/yagubogu/auth/gateway/FakeAuthGateway.java
@@ -1,4 +1,4 @@
-package com.yagubogu.auth.client;
+package com.yagubogu.auth.gateway;
 
 import com.yagubogu.auth.dto.AuthResponse;
 import com.yagubogu.auth.dto.GoogleAuthResponse;
